Conflicts:
	UPDATING

+++ sys/netproto/802_11/ieee80211_input.h
@@ -46,8 +46,8 @@
 #define	IEEE80211_VERIFY_LENGTH(_len, _minlen, _action) do {		\
 	if ((_len) < (_minlen)) {					\
 		IEEE80211_DISCARD(vap, IEEE80211_MSG_ELEMID,		\
-		    wh, NULL, "ie too short, got %d, expected %d",	\
-		    (_len), (_minlen));					\
+		    wh, NULL, "ie too short, got %jd, expected %jd",	\
+		    (intmax_t)(_len), (intmax_t)(_minlen));		\
 		vap->iv_stats.is_rx_elem_toosmall++;			\
 		_action;						\
 	}								\
